requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==4.2.2
django-environ==0.10.0
django-silk==5.0.3
djangorestframework==3.14.0
gprof2dot==2022.7.29
Pillow==10.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
